// src/types/partials/GroupCharacteristic.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_classes so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TQuantity } from '../partials/Quantity';
import { TRange } from '../partials/Range';
import { TReference } from '../partials/Reference';
import { TPeriod } from '../partials/Period';

export type TGroupCharacteristic = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    code: TCodeableConcept;
    valueCodeableConcept?: TCodeableConcept;
    valueBoolean?: Boolean;
    valueQuantity?: TQuantity;
    valueRange?: TRange;
    valueReference?: TReference;
    exclude: Boolean;
    period?: TPeriod;
};